refactor(web/asset): utilisation de l'AlertDialog général pour le dialogude de suppression
